# perf: cache folder nodes