--- a/macosx/FileListNode.mm
+++ b/macosx/FileListNode.mm
@@ -44,7 +44,6 @@
 - (void)insertChild:(FileListNode*)child
 {
     NSAssert(_isFolder, @"method can only be invoked on folders");
-
     [_children addObject:child];
 }