--- a/macosx/Torrent.mm
+++ b/macosx/Torrent.mm
@@ -1868,6 +1868,7 @@ bool trashDataFile(char const* filename, void* /*user_data*/, tr_error* error)
     {
         NSUInteger const count = self.fileCount;
         NSMutableArray* flatFileList = [NSMutableArray arrayWithCapacity:count];
+        NSMutableDictionary* folderCache = [NSMutableDictionary dictionary];
 
         FileListNode* tempNode = nil;
 
@@ -1894,7 +1895,8 @@ bool trashDataFile(char const* filename, void* /*user_data*/, tr_error* error)
                                 forParent:tempNode
                                  fileSize:file.length
                                     index:i
-                                 flatList:flatFileList];
+                                 flatList:flatFileList
+                              folderCache:folderCache];
         }
 
         [self sortFileList:tempNode.children];
@@ -1917,34 +1919,26 @@ bool trashDataFile(char const* filename, void* /*user_data*/, tr_error* error)
                        fileSize:(uint64_t)size
                           index:(NSInteger)index
                        flatList:(NSMutableArray<FileListNode*>*)flatFileList
+                    folderCache:(NSMutableDictionary<NSString*, FileListNode*>*)folderCache
 {
     NSParameterAssert(components.count > 0);
     NSParameterAssert(componentIndex < components.count);
 
     NSString* name = components[componentIndex];
     BOOL const isFolder = componentIndex < (components.count - 1);
+    NSString* path = [parent.path stringByAppendingPathComponent:parent.name];
 
     //determine if folder node already exists
-    __block FileListNode* node = nil;
-    if (isFolder)
-    {
-        [parent.children enumerateObjectsWithOptions:NSEnumerationConcurrent
-                                          usingBlock:^(FileListNode* searchNode, NSUInteger /*idx*/, BOOL* stop) {
-                                              if ([searchNode.name isEqualToString:name] && searchNode.isFolder)
-                                              {
-                                                  node = searchNode;
-                                                  *stop = YES;
-                                              }
-                                          }];
-    }
+    NSString* cacheKey = [path stringByAppendingPathComponent:name];
+    FileListNode* node = folderCache[cacheKey];
 
     //create new folder or file if it doesn't already exist
     if (!node)
     {
-        NSString* path = [parent.path stringByAppendingPathComponent:parent.name];
         if (isFolder)
         {
             node = [[FileListNode alloc] initWithFolderName:name path:path torrent:self];
+            folderCache[cacheKey] = node;
         }
         else
         {
@@ -1964,7 +1958,8 @@ bool trashDataFile(char const* filename, void* /*user_data*/, tr_error* error)
                             forParent:node
                              fileSize:size
                                 index:index
-                             flatList:flatFileList];
+                             flatList:flatFileList
+                          folderCache:folderCache];
     }
 }