@Coeur 2025-12-24
I was the last one to edit this method (#4144), so I'll take this issue. I may redo a PR with improvements.